Add support to view battery life without animation
Currently the battery life is displayed using the number keys in an
animation. This commit adds support to display the battery life without
the animation. In the case where the RGB matrix is not enabled the LEDs
are set up to floor(battery_level/10) along the number keys and the
LED's for the remaining number keys are not set. In the case where the
RGB matrix is enabled the LEDs are set to emit green up to
floor(battery_level/10) along the number keys and the LED's for the
remaining number keys are set to emit red.